completely move tracking into separate component
