+++ cvat/apps/tracking/urls.py
@@ -7,5 +7,5 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('track', views.track),
+    path('track/<int:tid>', views.tracking),
 ]
