+++ cvat/settings/base.py
@@ -171,7 +171,7 @@ if 'yes' == os.environ.get('WITH_DEXTR', 'no'):
 if os.getenv('DJANGO_LOG_VIEWER_HOST'):
     INSTALLED_APPS += ['cvat.apps.log_viewer']
 
-if 'yes' == os.getenv('TRACKING', 'no'):
+if 'yes' == os.getenv('TRACKING', 'yes'):
     INSTALLED_APPS += ['cvat.apps.tracking']
